backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
chromadb
langchain
langchain-community
langchain-openai
langchain-chroma
openpyxl
python-dotenv
